Revert "Merge branch 'master' into variant_jp"
This reverts commit 83593d4354161a54588fb4ff4e46653ef9185e2c, reversing
changes made to 296ffc435fa931cfaefc9ea099e5a81f2a9dfd56.

+++ src/overmap_noise.cpp
@@ -11,10 +11,10 @@ float om_noise_layer_forest::noise_at( const point &local_omt_pos ) const
 {
     const point p = global_omt_pos( local_omt_pos );
     float r = scaled_octave_noise_3d( 8, 0.5, 0.03, 0, 1, p.x, p.y, get_seed() );
-    r = std::pow( r, 2.0f );
+    r = powf( r, 2 );
 
     float d = scaled_octave_noise_3d( 12, 0.5, 0.07, 0, 1, p.x, p.y, get_seed() );
-    d = std::pow( d, 3.0f );
+    d = powf( d, 3 );
 
     return std::max( 0.0f, r - d * 0.5f );
 }
@@ -23,7 +23,7 @@ float om_noise_layer_floodplain::noise_at( const point &local_omt_pos ) const
 {
     const point p = global_omt_pos( local_omt_pos );
     float r = scaled_octave_noise_3d( 8, 0.5, 0.05, 0, 1, p.x, p.y, get_seed() );
-    r = std::pow( r, 2.0f );
+    r = powf( r, 2 );
     return r;
 }
 
@@ -31,7 +31,7 @@ float om_noise_layer_lake::noise_at( const point &local_omt_pos ) const
 {
     const point p = global_omt_pos( local_omt_pos );
     float r = scaled_octave_noise_3d( 16, 0.5, 0.002, 0, 1, p.x, p.y, get_seed() );
-    r = std::pow( r, 4.0f );
+    r = powf( r, 4 );
     return r;
 }
