+++ tests/list_test.cpp
@@ -1,8 +1,7 @@
 #include <algorithm> // std::find
 #include <cstdio> // log redirection
+#include <cstdlib> // abort
 #include <functional> // std::greater
-#include <initializer_list>
-#include <iterator>
 #include <utility> // std::move
 #include <vector> // range-insert testing
 
@@ -301,7 +300,7 @@ TEST_CASE( "list insert and erase", "[list]" )
     }
 
     SECTION( "erase randomly till half empty" ) {
-        size_t count = 0;
+        int count = 0;
         do {
             for( cata::list<int>::iterator it = test_list.begin(); it != test_list.end(); ) {
                 if( ( xor_rand() & 7 ) == 0 ) {
@@ -316,7 +315,7 @@ TEST_CASE( "list insert and erase", "[list]" )
 
         CHECK( test_list.size() == 500000 - count );
 
-        for( size_t i = 0; i < count; i++ ) {
+        for( int i = 0; i < count; i++ ) {
             test_list.push_front( 1 );
         }
 
@@ -432,7 +431,7 @@ TEST_CASE( "list insert and erase", "[list]" )
         CHECK( prev_capacity != test_list.capacity() );
         CHECK( test_list.capacity() == 1000 );
 
-        size_t count = 0;
+        int count = 0;
         for( int loop1 = 0; loop1 < 50000; loop1++ ) {
             for( int loop = 0; loop < 10; loop++ ) {
                 if( ( xor_rand() & 7 ) == 0 ) {
